+++ Demo/Demo/SwiftUIComponents/SwiftUICardDemo.swift
@@ -9,6 +9,8 @@ struct SwiftUICardDemo: View {
     @State private var cardNumberText: String = ""
     @State private var expirationDateText: String = ""
     @State private var cvvText: String = ""
+    @State private var vaultCustomerID: String = ""
+    @State var shouldVaultSelected = false
 
     @StateObject var baseViewModel = BaseViewModel()
 
@@ -32,6 +34,13 @@ struct SwiftUICardDemo: View {
                     .onChange(of: cvvText) { newValue in
                         cvvText = cardFormatter.formatFieldWith(newValue, field: .cvv)
                     }
+                HStack {
+                    Toggle("Should Vault", isOn: $shouldVaultSelected)
+                    Spacer()
+                }
+                if shouldVaultSelected {
+                    FloatingLabelTextField(placeholder: "Vault Customer ID (Optional)", text: $vaultCustomerID)
+                }
                 Button("\(DemoSettings.intent.rawValue.capitalized) Order") {
                     guard let card = baseViewModel.createCard(
                         cardNumber: cardNumberText,
@@ -46,7 +55,9 @@ struct SwiftUICardDemo: View {
                     Task {
                         await baseViewModel.checkoutWith(
                             card: card,
-                            orderID: orderID
+                            orderID: orderID,
+                            shouldVault: shouldVaultSelected,
+                            customerID: vaultCustomerID.isEmpty ? nil : vaultCustomerID
                         )
                     }
                 }
@@ -69,6 +80,6 @@ struct SwiftUICardDemo: View {
 struct ContentView_Previews: PreviewProvider {
 
     static var previews: some View {
-        SwiftUICardDemo(baseViewModel: BaseViewModel())
+        SwiftUICardDemo(shouldVaultSelected: true, baseViewModel: BaseViewModel())
     }
 }